pyproject.toml
[tool.poetry]
name = "microshop"
version = "0.1.0"
description = "Microshop to sell MacroItems!"
authors = ["Alexey Kogan <alekseikogan@yandex.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.1"}
uvicorn = {extras = ["standart"], version = "^0.30.3"}
pydantic-settings = "^2.3.4"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.31"
aiosqlite = "^0.20.0"
python-multipart = "^0.0.9"
alembic = "^1.13.2"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
